refactor(user): reorganize user module structure and update import paths
