+++ src/modules/user/application/queries/use-cases/list-user/list-user.query.ts
@@ -2,12 +2,35 @@ import { IQuery } from '@nestjs/cqrs';
 
 import { OrderTypeEnum } from '@common/domain/criteria/orders/order-type-enum';
 import { PrimitivesType } from '@common/domain/value-object/types/value-object';
-import { ListOrderByTypes } from '@user/infrastructure/controllers/user/queries/list-user/list-user-input.dto';
+
+export const USER_OPTIONS_ORDER_BY_VALUE = [
+  'status',
+  'username',
+  'email',
+  'name',
+  'lastName',
+  'secondLastName',
+  'phoneNumber',
+  'gender',
+  'street',
+  'extNumber',
+  'intNumber',
+  'neighborhood',
+  'zipCode',
+  'city',
+  'state',
+  'country',
+  'createdAt',
+  'updatedAt',
+  'archivedAt',
+] as const;
+
+export type USER_OPTIONS_ORDER_BY_VALUE_TYPE = typeof USER_OPTIONS_ORDER_BY_VALUE[number];
 
 export class ListUserQuery implements IQuery {
   public readonly orderType: OrderTypeEnum;
 
-  public readonly orderBy: ListOrderByTypes;
+  public readonly orderBy: USER_OPTIONS_ORDER_BY_VALUE_TYPE;
 
   public readonly limit: number;
