+++ src/modules/user/infrastructure/controllers/commands/archive-user/archive-user.controller.ts
@@ -4,8 +4,8 @@ import { MessagePattern, Payload } from '@nestjs/microservices';
 
 import { CMDS_HADES } from '@common/infrastructure/controllers/constants';
 import { ArchiveUserCommand } from '@user/application/commands/use-cases/archive-user/archive-user.command';
-import { ArchiveUserInputDto } from '@user/infrastructure/controllers/user/commands/archive-user/archive-user-input.dto';
-import { ArchiveUserOutputDto } from '@user/infrastructure/controllers/user/commands/archive-user/archive-user-output.dto';
+import { ArchiveUserInputDto } from '@user/infrastructure/controllers/commands/archive-user/archive-user-input.dto';
+import { ArchiveUserOutputDto } from '@user/infrastructure/controllers/commands/archive-user/archive-user-output.dto';
 
 @Controller()
 export class ArchiveUserController {
