+++ src/modules/user/infrastructure/controllers/commands/create-user/create-user-input.dto.ts
@@ -1,8 +1,8 @@
 import { Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 
-import { AccountDTO } from '@user/infrastructure/controllers/user/commands/create-user/dtos/account.dto';
-import { ProfileDTO } from '@user/infrastructure/controllers/user/commands/create-user/dtos/profile.dto';
+import { AccountDTO } from '@user/infrastructure/controllers/commands/create-user/dtos/account.dto';
+import { ProfileDTO } from '@user/infrastructure/controllers/commands/create-user/dtos/profile.dto';
 
 export class CreateUserInput {
   @ValidateNested({ each: true })
