+++ src/modules/user/infrastructure/controllers/commands/create-user/create-user-output.dto.ts
