+++ src/modules/user/infrastructure/controllers/commands/destroy-user/destroy-user-input.dto.ts
