+++ src/modules/user/infrastructure/controllers/commands/destroy-user/destroy-user-output.dto.ts
