+++ src/modules/user/infrastructure/controllers/commands/update-user/dtos/profile.dto.ts
@@ -6,8 +6,8 @@ import {
   Length,
   IsEnum,
   ValidateNested,
-  ArrayNotEmpty,
   IsArray,
+  ArrayNotEmpty,
 } from 'class-validator';
 
 import {
@@ -21,36 +21,41 @@ import {
   MIN_PROFILE_PHONE_NUMBER_LENGTH,
   ProfileGenderEnum,
 } from '@user/domain/constants/general-rules';
-import { AddressDTO } from '@user/infrastructure/controllers/user/commands/create-user/dtos/address.dto';
+import { AddressDTO } from '@user/infrastructure/controllers/commands/update-user/dtos/address.dto';
 
 export class ProfileDTO {
+  @IsOptional()
   @IsArray()
   @ArrayNotEmpty()
   @IsString({ each: true })
   @Length(MIN_PROFILE_NAME_LENGTH, MAX_PROFILE_NAME_LENGTH, { each: true })
-  public readonly names: string[];
+  public readonly names?: string[];
 
+  @IsOptional()
   @IsString()
   @IsNotEmpty()
   @Length(MIN_PROFILE_LAST_NAME_LENGTH, MAX_PROFILE_LAST_NAME_LENGTH)
-  public readonly lastName: string;
+  public readonly lastName?: string;
 
-  @IsString()
   @IsOptional()
+  @IsString()
   @Length(MIN_PROFILE_SECOND_LAST_NAME_LENGTH, MAX_PROFILE_SECOND_LAST_NAME_LENGTH)
   public readonly secondLastName?: string;
 
+  @IsOptional()
   @IsString()
   @IsNotEmpty()
   @Length(MIN_PROFILE_PHONE_NUMBER_LENGTH, MAX_PROFILE_PHONE_NUMBER_LENGTH)
-  public readonly phoneNumber: string;
+  public readonly phoneNumber?: string;
 
+  @IsOptional()
   @IsEnum(ProfileGenderEnum)
   @IsNotEmpty()
-  public readonly gender: ProfileGenderEnum;
+  public readonly gender?: ProfileGenderEnum;
 
+  @IsOptional()
   @IsNotEmpty()
   @ValidateNested()
   @Type(() => AddressDTO)
-  public readonly address: AddressDTO;
+  public readonly address?: AddressDTO;
 }
