+++ src/modules/user/infrastructure/controllers/commands/update-user/update-user-input.dto.ts
@@ -1,8 +1,8 @@
 import { Type } from 'class-transformer';
 import { IsOptional, IsUUID, ValidateNested } from 'class-validator';
 
-import { AccountDTO } from '@user/infrastructure/controllers/user/commands/update-user/dtos/account.dto';
-import { ProfileDTO } from '@user/infrastructure/controllers/user/commands/update-user/dtos/profile.dto';
+import { AccountDTO } from '@user/infrastructure/controllers/commands/update-user/dtos/account.dto';
+import { ProfileDTO } from '@user/infrastructure/controllers/commands/update-user/dtos/profile.dto';
 
 export class UpdateUserInput {
   @IsUUID()
