+++ src/modules/user/infrastructure/controllers/queries/get-user/get-user-output.dto.ts
