+++ src/modules/user/infrastructure/framework/user.module.ts
@@ -14,12 +14,12 @@ import { DestroyUserDomainService } from '@user/domain/domain-services/destroy-u
 import { GetUserDomainService } from '@user/domain/domain-services/get-user.domain-service';
 import { ListUserDomainService } from '@user/domain/domain-services/list-user.domain-service';
 import { UpdateUserDomainService } from '@user/domain/domain-services/update-user.domain-service';
-import { ArchiveUserController } from '@user/infrastructure/controllers/user/commands/archive-user/archive-user.controller';
-import { CreateUserController } from '@user/infrastructure/controllers/user/commands/create-user/create-user.controller';
-import { DestroyUserController } from '@user/infrastructure/controllers/user/commands/destroy-user/destroy-user.controller';
-import { UpdateUserController } from '@user/infrastructure/controllers/user/commands/update-user/update-user.controller';
-import { GetUserController } from '@user/infrastructure/controllers/user/queries/get-user/get-user.controller';
-import { ListUserController } from '@user/infrastructure/controllers/user/queries/list-user/list-user.controller';
+import { ArchiveUserController } from '@user/infrastructure/controllers/commands/archive-user/archive-user.controller';
+import { CreateUserController } from '@user/infrastructure/controllers/commands/create-user/create-user.controller';
+import { DestroyUserController } from '@user/infrastructure/controllers/commands/destroy-user/destroy-user.controller';
+import { UpdateUserController } from '@user/infrastructure/controllers/commands/update-user/update-user.controller';
+import { GetUserController } from '@user/infrastructure/controllers/queries/get-user/get-user.controller';
+import { ListUserController } from '@user/infrastructure/controllers/queries/list-user/list-user.controller';
 import { AccountEntity } from '@user/infrastructure/persistence/typeorm/entities/account.entity';
 import { ProfileEntity } from '@user/infrastructure/persistence/typeorm/entities/profile.entity';
 import { SessionEntity } from '@session/infrastructure/persistence/typeorm/entities/session.entity';
